split bundle and PE concerns
