+++ syft/pkg/cataloger/internal/dotnet/pe/bundle_test.go
