+++ syft/pkg/cataloger/internal/dotnet/pe/pe.go
