+++ syft/pkg/cataloger/internal/dotnet/pe/pe_test.go
