+++ syft/pkg/cataloger/internal/dotnet/pe/testdata/Makefile
