+++ syft/pkg/cataloger/internal/dotnet/pe/testdata/image-dotnet31-single-file/src/Program.cs
