+++ syft/pkg/cataloger/internal/dotnet/pe/testdata/image-dotnet31-single-file/src/hello.csproj
