+++ syft/pkg/cataloger/internal/dotnet/pe/testdata/image-dotnet5-single-file/src/Program.cs
