+++ syft/pkg/cataloger/internal/dotnet/pe/testdata/image-dotnet6-single-file/Dockerfile
@@ -1,6 +1,6 @@
-FROM --platform=linux/amd64 mcr.microsoft.com/dotnet/sdk:5.0 AS build
+FROM --platform=linux/amd64 mcr.microsoft.com/dotnet/sdk:6.0 AS build
 WORKDIR /src
-COPY src/ .
+COPY src .
 RUN dotnet publish -c Release -r win-x64 \
     -p:PublishSingleFile=true \
     -p:SelfContained=true \
