+++ syft/pkg/cataloger/internal/dotnet/pe/testdata/image-dotnet6-single-file/src/Program.cs
