+++ syft/pkg/cataloger/internal/dotnet/pe/testdata/image-net8-app
@@ -1,0 +1,1 @@
+../../../../dotnet/testdata/image-net8-app
