+++ syft/pkg/cataloger/internal/dotnet/pe/testdata/image-net8-app-single-file
@@ -1,0 +1,1 @@
+../../../../dotnet/testdata/image-net8-app-single-file
