+++ syft/pkg/cataloger/internal/dotnet/pe/testdata/net8-app-single-file.deps.json
