--- syft/pkg/cataloger/internal/pe/testdata/image-net8-app-single-file
@@ -1,1 +1,0 @@
-../../../dotnet/testdata/image-net8-app-single-file
